animations, palettes, and textures for world objects (no data yet, still debugging my exports) (#208)

+++ Source/ACE/Entity/ModelData.cs
@@ -18,7 +18,7 @@ namespace ACE.Entity
 
         private List<Model> models = new List<Model>();
 
-        public void AddPalette(ushort paletteID, byte offset, byte length)
+        public void AddPalette(uint paletteID, ushort offset, ushort length)
         {
             ModelPalette newpalette = new ModelPalette(paletteID, offset, length);
             modelPalettes.Add(newpalette);
